The program now uses has a main script and modules used by the main script
